Relocate CompilerInstanceWithContext to DependencyScanningWorker.cpp and hide it as an implementation detail.
